+++ nodalia-cards-2.0.0-alpha.58.js
@@ -1,2 +1,2 @@
-import "./nodalia-cards-2.0.0-alpha.59.js";
-if(typeof window!=="undefined"){window.__NODALIA_LOADER__={"mode":"compat","pkgVersion":"2.0.0-alpha.59","contentSha256_12":"59f4e4b8027e","file":"nodalia-cards-2.0.0-alpha.58.js","targetFile":"nodalia-cards-2.0.0-alpha.59.js","fallbackFile":"nodalia-cards.js"};}
+import "./nodalia-cards-2.0.0-alpha.60.js";
+if(typeof window!=="undefined"){window.__NODALIA_LOADER__={"mode":"compat","pkgVersion":"2.0.0-alpha.60","contentSha256_12":"09a886320889","file":"nodalia-cards-2.0.0-alpha.58.js","targetFile":"nodalia-cards-2.0.0-alpha.60.js","fallbackFile":"nodalia-cards.js"};}