+++ nodalia-cards.manifest.js
@@ -1,23 +1,23 @@
 export default {
-  "pkgVersion": "2.0.0-alpha.59",
-  "contentSha256_12": "59f4e4b8027e",
+  "pkgVersion": "2.0.0-alpha.60",
+  "contentSha256_12": "09a886320889",
   "file": "nodalia-cards.bundle.js",
   "loaderFile": "nodalia-cards.js",
   "hacsFile": "nodalia-cards.js",
   "compatLoaderFiles": [
-    "nodalia-cards-2.0.0-alpha.57.js",
-    "nodalia-cards-2.0.0-alpha.58.js"
+    "nodalia-cards-2.0.0-alpha.58.js",
+    "nodalia-cards-2.0.0-alpha.59.js"
   ],
-  "splitCoreFile": "nodalia-cards-core-2.0.0-alpha.59.js",
+  "splitCoreFile": "nodalia-cards-core-2.0.0-alpha.60.js",
   "splitCoreSha256_12": "ddd1431fcb0d",
-  "splitSuiteFile": "nodalia-cards-suite-2.0.0-alpha.59.js",
-  "splitSuiteSha256_12": "a5c80d88e21c",
-  "editorFile": "nodalia-cards-editor-2.0.0-alpha.59.js",
+  "splitSuiteFile": "nodalia-cards-suite-2.0.0-alpha.60.js",
+  "splitSuiteSha256_12": "edcbb7221661",
+  "editorFile": "nodalia-cards-editor-2.0.0-alpha.60.js",
   "editorSha256_12": "d1c3a2e97124"
 };
-export const pkgVersion = "2.0.0-alpha.59";
-export const contentSha256_12 = "59f4e4b8027e";
+export const pkgVersion = "2.0.0-alpha.60";
+export const contentSha256_12 = "09a886320889";
 export const file = "nodalia-cards.bundle.js";
-export const splitCoreFile = "nodalia-cards-core-2.0.0-alpha.59.js";
-export const splitSuiteFile = "nodalia-cards-suite-2.0.0-alpha.59.js";
-export const editorFile = "nodalia-cards-editor-2.0.0-alpha.59.js";
+export const splitCoreFile = "nodalia-cards-core-2.0.0-alpha.60.js";
+export const splitSuiteFile = "nodalia-cards-suite-2.0.0-alpha.60.js";
+export const editorFile = "nodalia-cards-editor-2.0.0-alpha.60.js";